requirements.txt
# LLM Libraries
google-genai>=1.16.0

# Google API Client (for Custom Search)
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.7.0

# Web Framework & Server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
tenacity>=8.2.3

# Web Scraping
beautifulsoup4>=4.9.3
requests>=2.25.1
readability-lxml>=0.8.1
# newspaper3k>=0.2.8

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utilities
requests>=2.25.0